Mise a jour de la generation des nombres aleatoires pour toutes les simulations. Jai fait un essai, il faut que je repasse sous windows pour verifier les resultats et valider la nouvelle methode.

+++ trunk/ projet-ia/inc/meteo.h
@@ -19,6 +19,6 @@
 float SimulationTemp(ST_SimuMeteo Meteo_Precedente,ST_DonneeMeteo Normales_Mensuelles);
 void SimulationConditionsMeteo(ST_DonneeMeteo Normales_Mensuelles, PTR_ST_SimuMeteo Meteo_Jour);
 void SimulationHeuresEnsoleillement(ST_DonneeMeteo NormalesMensuelles, PTR_ST_SimuMeteo Meteo_Jour);
-int my_rand(void);
+float my_rand(void);
 
 #endif
